Fixed error while fetching keys
The first call of getKeysPaged resulted in a fetch of the latest
keys always as the call did not take into account the provided
at from which block to fetch.

Furthermore, this resulted in an error of vestings being migrated
for accounts, that do not exist. For this reason, the verification
now checks if the account exists and if the balances are correct.