Fixed storage fetch, added delta to vesting
This commit fixes that the call getStorage() did not use the
at-option. Meaning, that although the keys where fetch at a specific
block, the storage was fetched at the latest. Now, the storage is
fetched at the block where the keys are fetched from.

Furthermore, this commit ensures, that the vestings are vested up
to the point when the migration is run. To ensure this, the script
does fetch the latest head and calculates the new-vestings as if
everybody called vest at this latest block.